Fix cross-platform CI: Windows compilation, macOS test, Linux CMake
Windows (MSVC) fixes:
- GLFWDynamicLibrary.cpp: change #ifdef ENGINE_PLATFORM_* to #if
  (Platform.h defines macros to 0 on non-matching platforms, so
  #ifdef was always true, compiling POSIX dlopen on Windows)
- StackTrace.cpp: change DWORD to DWORD64 for SymFromAddr displacement
  (MSVC requires PDWORD64 = DWORD64*, not DWORD*)
- Path.cpp: disambiguate Path(std::string) by using Path(string_view)
  (std::string is implicitly convertible to both filesystem::path
  and string_view, causing ambiguous overload resolution)

macOS (Clang) fixes:
- Instrumentation.cpp: add missing 'using engine::core::u64'
  (pthread_threadid_np requires unsigned 64-bit, not signed i64)
- Timing.cpp: Clock::Timestamp() now uses system_clock instead of
  high_resolution_clock (which is steady_clock on macOS, returning
  time-since-boot instead of Unix epoch timestamp)

CI infrastructure:
- ci.yml: ubuntu-24.04 with GCC 13/14 (gcc-13 not in ubuntu-22.04)
- patch_glfw.cmake: cross-platform CMake -P script to force-disable
  GLFW tests/examples/docs and inject X11 compat on Linux
- Dependencies.cmake: yaml-cpp FORCE cache vars, GLAD generation
  with system_clock (no --reproducible), GLFW patch_command
- CMakeLists.txt: add C to LANGUAGES (CMake 4.x requires explicit
  C language declaration for .c files like glad.c)
- XInput.h compat stub for GLFW's X11 header detection

+++ Engine/Source/Diagnostics/StackTrace.cpp
@@ -237,7 +237,7 @@ std::string StackTrace::Capture(usize maxFrames)
     symbol->SizeOfStruct = sizeof(SYMBOL_INFO);
     symbol->MaxNameLen   = static_cast<ULONG>(kMaxNameLen);
 
-    DWORD displacement = 0;
+    DWORD64 displacement = 0;
     IMAGEHLP_LINE64 line = {};
     line.SizeOfStruct = sizeof(IMAGEHLP_LINE64);
